requirements.txt
numpy
pandas
scikit-learn
joblib
pytest>=8
torch
fair-esm  # optional, required for --method esm
